fix macro related test failures
for some reason with `codes` used within some of the other macros, in
doctests where nesting a pair of those other macros, e.g.
`seq!(rgb_fg!(...))` this would result in an error from `cargo test` which
would complain that `codes` could not be found within `$crate`.

i could not find anything i'd done wrong; i presume it to be a bug or
limitation in the macro system.

removing use of `codes` from within the other macros here resolves the
failure.