ci: try to fix failures for macos
okay, the previous commit fixed things for Linux, but the macos ones fail
because the `-e` param does not exist for the macOS version of the `script`
command. printing the man page (because there's not an easily available
copy online) using a previous version of this commit, i can see that the
param order also differs and the command string seems to not be a string
param of a `-c`/`--command` option but just all remaining arguments on the
end.